refactor(agent): erase the tool surface
